File naming changed to index for better routing. UT fixed for PR release
