+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "my-pets",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "author": "Nicolás Omar González Passerino",
   "license": "MIT",
   "private": false,
