+++ src/components/app/app.test.js
@@ -3,7 +3,7 @@ import { MockedProvider } from '@apollo/client/testing'
 import { Provider } from 'react-redux'
 import { createStore } from 'redux'
 // REDUX
-import store from '../../../redux/reducers'
+import store from '../../redux/reducers'
 // COMPONENTS
 import App from './app'
