+++ src/components/molecules/grid-layout/index.test.js
@@ -1,11 +1,11 @@
 import { render, screen } from '@testing-library/react'
-import GridLayout from './grid-layout'
+import GridLayout from '.'
 
 describe('[GridLayout]', () => {
   describe('[HAPPY PATH]', () => {
     test('Renders with required properties only', () => {
-      render(<GridLayout header={{ title: 'Test' }} />)
-      const element = screen.getByText('Test')
+      render(<GridLayout />)
+      const element = screen.getByTestId('grid-test-layout')
       expect(element).toBeInTheDocument()
     })
   })
