+++ src/components/pages/home-page/index.test.js
@@ -1,14 +1,14 @@
 import { render, screen } from '@testing-library/react'
 // COMPONENTS
-import Home from '../../molecules/home/home'
+import HomePage from '.'
 // HELPER FUNCTIONS
 import { setLoggedUser } from '../../../functions/local-storage'
 
-describe('[Home]', () => {
+describe('[HomePage]', () => {
   describe('[HAPPY PATH]', () => {
     test('Renders with a dummy logged User', () => {
       setLoggedUser({ name: 'Test' })
-      render(<Home />)
+      render(<HomePage />)
       const element = screen.getByText(`HELLO TEST`)
       expect(element).toBeInTheDocument()
     })
