+++ src/components/pages/update-user-page/index.jsx
@@ -5,7 +5,7 @@ import { useHistory } from 'react-router'
 import { useMutation } from '@apollo/client'
 import { UPDATE_USER } from '../../../graphql/mutations'
 // COMPONENTS
-import FormTemplate from '../../templates/form-template/form-template'
+import FormTemplate from '../../templates/form-template'
 // FORM CONFIG
 import { inputs, header, saveButton, goBackButton } from './update-user.config.json'
 // CONSTANTS
