+++ src/components/templates/form-template/index.jsx
@@ -1,8 +1,8 @@
 import { array, bool, func, object } from 'prop-types'
 import React from 'react'
 // COMPONENTS
-import BasicFrame from '../../organisms/basic-frame/basic-frame'
-import Form from '../../organisms/form/form'
+import BasicFrame from '../../organisms/basic-frame'
+import Form from '../../organisms/form'
 
 const FormTemplate = ({
   header,
